Allow access to the instance through subdomains of glitch.cat.family
Rails 6 forces you to whitelist what domain names an application
may be accessed from. while it is possible to supply a regular
expression, it is not clear how to do this when mastodon parses the
environment variable you supply the regex from. meanwhile, disabling
whitelisting entirely DOES work

this is theoretically less secure; if we can find a way to supply a
regex, that would be preferable.

	--- yr cat