[UNTESTED] Require a word boundary after a mention
It is very longstanding behavior for `MENTION_RE` to match strings of
the form `@user@domain@thoasuhsantoshu`. This change requires Unicode
whitespace after a message, with an edge case of allowing a mention to
occur at the very end of the post.

This is really a just a start when it comes to mention matching. The
new test is as well. As well as the entire service spec.